requirements.txt
python-dotenv~=1.1.0
psycopg2-binary~=2.9.10
pandas~=2.3.0
plotly~=6.1.2
flask~=3.1.1
asyncpraw~=7.8.1
websockets~=10.4
nltk~=3.9.1
aiohttp~=3.12.11
numpy==1.26.4
aiohttp
asyncpraw
nltk
alpaca-trade-api
torch
transformers
pandas-ta